requirements.txt
pytest
python-decouple
requests